[チュートリアル]Post単体ページをTypeScriptに変換
